[PAIG-2210] Fixed cyclic dependency issue
