+++ paig-server/backend/paig/api/guardrails/transformers/__init__.py
@@ -1,7 +1,6 @@
 from abc import ABC
 
 from api.guardrails.providers import GuardrailConfig
-from api.guardrails.transformers.backend.aws_gr_transformer import AWSGuardrailTransformer
 
 
 class GuardrailTransformer(ABC):
@@ -32,15 +31,3 @@ class GuardrailTransformer(ABC):
             list[GuardrailConfig]: The transformed guardrail configurations.
         """
         pass
-
-
-class GuardrailTransformerFactory:
-    @staticmethod
-    def get_transformer(provider: str) -> GuardrailTransformer:
-        transformers = {
-            "AWS": AWSGuardrailTransformer
-        }
-        transformer_class = transformers.get(provider)
-        if not transformer_class:
-            raise ValueError(f"Unsupported provider: {provider}")
-        return transformer_class()
